# +faster C code  +Rust