+_C2.c   +Rust
+ fordítás a measure.sh szkriptben + C2 + Rust

Megjegyzés:

time ./proc_speed_test_C2

real 0m20,167s
user 0m2,818s   <---- programozási nyelv futási ideje
sys 0m17,079s   < --- kernel "nyaggatás" ideje

$ time pypy proc_speed_test.py

real 0m32,157s
user 0m9,461s
sys 0m22,287s

Ebből az látszik, hogy a kernelhívás visz el sok időt, maga a progam futásideje kevésbé jelentős. Egyébként az Intel Meltdown hibájának a javítása a sok kernelhívás miatt sokat rontott ezen a teszten.