@haa-zee 2018-03-31
Szia! Köszi szépen, de azt hiszem, ehhez kellene egy kis segítség: a measure.sh-t nem szeretném piszkálni, az maradjon meg kizárólag a tesztek futtatására, a fordítást inkább kiraknám egy Makefile-ba. Viszont fogalmam sincs, hogyan tudnám fájlonként beolvasztani az általad javasolt változtatásokat úgy, hogy a measure.sh kimaradjon. Tudnál ebben segíteni?

A kernel témával kapcsolatban: akkor félreértettem a dolgot, azt hittem, valami olyan hibát vétettem, ami miatt az indokoltnál több időt tölt a kernelben valamelyik program. Ez így, a magam szempontjait figyelembe véve nem izgalmas, mivel ezt nem tudom befolyásolni. A teszt meg csak arra készült, hogy lássam: szabad-e pythonban monitoring applikációt írni, mennyire terhelné egy pythonos cucc jobban a rendszert, mintha valami más nyelven íródott volna.