@hg2ecz 2018-03-31
Legegyszerűbben: másold ki magadnak a jelenlegi measure.sh-t, majd olvasszad be a pull req-t és végül írd felül a nálad levő eredetivel aztán commit & push.

Egyébként szerintem elfér benne a fordító rész, nem időigényes, ellenben sokkal könnyebben használható az egész. Ezért raktam bele. Próbáld ki.

Tempókról: onnantól kezdve lesz igazán jelentős a különbség, ha nem rendszerhívásról szól szinte az egész kód, hanem a kódon belül is számolgat. Ezért érdemes kíírni a külön a sys és user időparamétereket.