@haa-zee 2018-03-31
Ilyen alapon a gittel is visszahozathatom a szkript előző verzióját. Azt hittem, csak én nem találom itt a githubon, hogy egyesével fogadjam/utasítsam el a requesteket. Jelen esetben így egyszerűbb klónozni a forkodat és abból áttenni a te részeidet ebbe.

A measure.sh kapcsán: elférni elfér benne a fordítás, de nem oda való. Ha úgy vesszük, a measure.sh "az üzemeltetés" része, a fordítás meg a fejlesztési oldalé.

Sys és user idő ott van egyébként minden '=' kezdetű sorban: név, real time; system cpu; user cpu; stb. Lásd még man time! Csak esetemben ez nem bírt igazán jelentőséggel, mert arra nincs befolyásom és csak a kiolvasott egy-egy sorból kellett volna pár numerikus értéket tárolni, bár az is látszik, hogy ha nem is arányosan, de azért a hosszabb futásidejű kódok System-ben is több időt töltenek.

Majd még az outputon kell egy kicsit finomítani, hogy egyszerűbb legyen a .csv -> libreoffice (calc) konverzió és az azt követő tábla készítés :)