@haa-zee 2018-03-31
Köszönettel átvettem a programjaidat, a measure.sh-t békénhagytam (azt leszámítva, hogy a te programjaid is belekerültek)